requirements.txt
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.1
amqp==5.2.0
anyio==4.3.0
apache-airflow==2.9.1
apache-airflow-providers-celery==3.7.0
apache-airflow-providers-cncf-kubernetes==8.2.0
apache-airflow-providers-common-io==1.3.1
apache-airflow-providers-common-sql==1.13.0
apache-airflow-providers-fab==1.1.0
apache-airflow-providers-ftp==3.9.0
apache-airflow-providers-http==4.11.0
apache-airflow-providers-imap==3.6.0
apache-airflow-providers-postgres==5.11.0
apache-airflow-providers-smtp==1.7.0
apache-airflow-providers-sqlite==3.8.0
apispec==6.6.1
argcomplete==3.3.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
Babel==2.15.0
billiard==4.2.0
blinker==1.8.2
cachelib==0.9.0
cachetools==5.3.3
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
clickclick==20.10.2
colorama==0.4.6
colorlog==4.8.0
ConfigUpdater==3.2
connexion==2.14.2
cron-descriptor==1.4.3
croniter==2.0.5
cryptography==42.0.7
Deprecated==1.2.14
dill==0.3.8
distlib==0.3.8
dnspython==2.6.1
docutils==0.21.2
email_validator==2.1.1
exceptiongroup==1.2.1
filelock==3.14.0
Flask==2.2.5
Flask-AppBuilder==4.4.1
Flask-Babel==2.0.0
Flask-Caching==2.3.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.6.0
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1
flower==2.0.1
frozenlist==1.4.1
fsspec==2024.5.0
google-auth==2.29.0
google-re2==1.1.20240501
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.63.0
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
humanize==4.9.0
idna==3.7
importlib-metadata==7.0.0
importlib_resources==6.4.0
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kombu==5.3.7
kubernetes==29.0.0
kubernetes_asyncio==29.0.0
lazy-object-proxy==1.10.0
limits==3.12.0
linkify-it-py==2.0.3
lockfile==0.12.2
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
marshmallow-oneofschema==3.1.1
marshmallow-sqlalchemy==0.28.2
mdit-py-plugins==0.4.1
mdurl==0.1.2
methodtools==0.4.7
more-itertools==10.2.0
multidict==6.0.5
numpy==1.26.4
oauthlib==3.2.2
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-exporter-otlp-proto-http==1.24.0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
ordered-set==4.1.0
packaging==24.0
pandas==2.2.2
pathspec==0.12.1
pendulum==3.0.0
platformdirs==4.2.2
pluggy==1.5.0
prison==0.2.1
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
Pygments==2.18.0
PyJWT==2.8.0
python-daemon==3.0.1
python-dateutil==2.9.0.post0
python-nvd3==0.16.0
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
redis==5.0.4
referencing==0.35.1
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.7.1
rich-argparse==1.4.0
rpds-py==0.18.1
rsa==4.9
setproctitle==1.3.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==1.4.52
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.0
tabulate==0.9.0
tenacity==8.3.0
termcolor==2.4.0
text-unidecode==1.3
time-machine==2.14.1
tornado==6.4
typing_extensions==4.11.0
tzdata==2024.1
uc-micro-py==1.0.3
unicodecsv==0.14.1
universal_pathlib==0.2.2
urllib3==2.2.1
vine==5.1.0
virtualenv==20.26.2
wcwidth==0.2.13
websocket-client==1.8.0
Werkzeug==2.2.3
wirerope==0.4.7
wrapt==1.16.0
WTForms==3.1.2
yarl==1.9.4
zipp==3.18.2
